+++ packages/charts/chart-utilities/src/index.ts
@@ -114,3 +114,5 @@ export {
 } from './PlotlySchemaConverter';
 
 export { decodeBase64Fields } from './DecodeBase64Data';
+
+export { convertToLocaleString, handleFloatingPointPrecisionError } from './formatter';
