+++ packages/charts/react-charts/library/src/utilities/UtilityUnitTests.test.ts
@@ -12,7 +12,7 @@ import { ScaleBand } from 'd3-scale';
 import { select as d3Select } from 'd3-selection';
 import { conditionalDescribe, isTimezoneSet } from './TestUtility.test';
 import * as vbcUtils from './vbc-utils';
-import { convertToLocaleString } from './locale-util';
+import { convertToLocaleString } from '@fluentui/chart-utilities';
 const { Timezone } = require('../../scripts/constants');
 const env = require('../../config/tests');
